Tests: Split DOMConveter tests.
